refector the losses
