+++ examples/train_vae.py
@@ -9,11 +9,10 @@ import tensorflow_probability as tfp
 from galcheat.utilities import mean_sky_level
 
 from maddeb.batch_generator import COSMOSsequence
-from maddeb.FlowVAEnet import (
-    FlowVAEnet,
-    deblender_loss_fn_wrapper,
-    deblender_ssim_loss_fn_wrapper,
-)
+from maddeb.FlowVAEnet import FlowVAEnet
+   
+from maddeb.losses import deblender_loss_fn_wrapper, deblender_ssim_loss_fn_wrapper
+
 from maddeb.callbacks import define_callbacks
 from maddeb.utils import get_data_dir_path, listdir_fullpath
